Separate senior and assistant pages
