+++ app/api/auth/[...nextauth]/route.ts
@@ -1,96 +1,6 @@
-import { Role } from "backend/role";
-import { callTabidoo } from "backend/tabidoo";
-import { AssistantResponse } from "backend/tabidoo/interfaces/assistant";
-import { SeniorResponse } from "backend/tabidoo/interfaces/senior";
-import { getFullName } from "backend/utils/getFullName";
-import { verifyPassword } from "helper/auth";
+import { authOptions } from "app/lib/auth";
 import NextAuth from "next-auth/next";
-import CredentialsProvider from "next-auth/providers/credentials";
 
-const handler = NextAuth({
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user?.role) {
-        token.role = user.role;
-      }
-      return token;
-    },
-    session({ session, token }) {
-      /* Step 2: update the session.user based on the token object */
-      if (token && session.user) {
-        session.user.role = token.role;
-      }
-      return session;
-    },
-  },
-  providers: [
-    CredentialsProvider({
-      credentials: {},
-      async authorize(credentials) {
-        const { email, password } = credentials as {
-          email: string;
-          password: string;
-        };
-
-        if (!email) {
-          throw Error("Zadejte prosím email");
-        }
-
-        if (!password) {
-          throw Error("Zadejte prosím heslo");
-        }
-
-        const tabidooRequestPayload = {
-          method: "POST",
-          body: {
-            filter: [
-              {
-                field: "email",
-                operator: "eq",
-                value: email,
-              },
-            ],
-          },
-          urlParams: {
-            limit: "1",
-          },
-        };
-
-        const [foundSeniors, foundAssistants] = await Promise.all([
-          callTabidoo<SeniorResponse[]>(
-            "/tables/senior/data/filter",
-            tabidooRequestPayload
-          ),
-          callTabidoo<AssistantResponse[]>(
-            "/tables/uzivatel/data/filter",
-            tabidooRequestPayload
-          ),
-        ]);
-
-        const foundUsers = [...foundSeniors, ...foundAssistants];
-
-        if (foundUsers.length === 0) {
-          throw new Error("Uživatel nebyl nalezen");
-        }
-
-        const user = foundUsers[0];
-
-        const isValid = await verifyPassword(password, user.fields.heslo);
-
-        if (!isValid) {
-          throw Error("Špatně zadaný email nebo heslo");
-        }
-
-        return {
-          id: user.id,
-          email: user.fields.email,
-          role: foundSeniors.length > 0 ? Role.SENIOR : Role.DA,
-          name: getFullName(user),
-        };
-      },
-    }),
-  ],
-  secret: process.env.NEXTAUTH_SECRET,
-});
+const handler = NextAuth(authOptions);
 
 export { handler as GET, handler as POST };
