+++ app/senior/page.tsx
@@ -1,0 +1,15 @@
+import Profile from "components/senior/profile";
+import Head from "next/head";
+
+function Page() {
+  return (
+    <>
+      <Head>
+        <title>Profil Seniora | Moudrá Síť App</title>
+      </Head>
+      <Profile />
+    </>
+  );
+}
+
+export default Page;
